This is sunday work 29/april /2017 created a child componenet which is reusable, created which accepts input, learned to declare application directives in angular module. listed in decorations sections, Learned to bind a parent componenet to child component. later on sunday 30/ april 2017
created a service class that can be shared by many component
used ng OnInit lifecylce to hook to get the hero data when the
component activates.
define hero service as a PROVIDER for the app component. created a mock
hero data and imported them into the service.
Designed the service to return a Promise and the component to get data
from the promise.